- Created missing endpoints, no business logic created yet. - Removed ObjectManager from constructor - Renamed controllers to be in line with Magento controllers
